+++ src/main/java/mada/App.java
@@ -4,9 +4,6 @@ import java.io.File;
 import java.math.BigInteger;
 import java.util.List;
 
-import org.junit.Assert;
-import org.junit.Test;
-
 public class App {
 	private static final String CHIFFRE_TEXT_FILE_NAME = "/chiffre.txt";
 	private static final String PRIVATE_KEY_FILE_NAME = "/sk.txt";
@@ -57,7 +54,7 @@ public class App {
 		return key;
 	}
 
-	private String encrypt(Tuple tuple, String plainText) {
+	public String encrypt(Tuple tuple, String plainText) {
 		Encryptor encryptor = new Encryptor();
 		String chiffreText = encryptor.encrypt(tuple, plainText);
 
@@ -66,22 +63,9 @@ public class App {
 		return chiffreText;
 	}
 
-	private String decrypt(Tuple privateKey, String chiffreText) {
+	public String decrypt(Tuple privateKey, String chiffreText) {
 		Decryptor decryptor = new Decryptor();
 		String plainText = decryptor.decrypt(privateKey, chiffreText);
 		return plainText;
 	}
-
-	@Test
-	public void testWithGivenValues() {
-		App app = new App();
-		BigInteger first = new BigInteger(
-				"16803438598030752881652661943729421682719289436596914422084556525786143242587823357440464415199237062279518632303253933220399688704185812346878881794965528739303640123538737959915328042133298634259748307946908213504142640812530197351030910115816579697756196422484983661518333751973651657437794766379852236386712862452160121063539921560702971674741294851536681666477917939238975289430870693260123741653964788745826206110446743767804057776626047201196295322367520161600144776544964528374283143367344047873033797568502984831163759801628072488471869900411279509552067591214099783281874706197386478061529452599356907098203");
-		BigInteger second = new BigInteger(
-				"12552902182293782601169303997788253294539781978521267882125178188595202511769797881455173372350572155151181053739643397112128238388516916958219680086053943902643802993855884031608016015606514912716831979016016589483029732083260843982927542126741595212207541093556957368879809911391702572539130475413753843496717359797858011974688549745704258562157053465493269685009560932033821844711601832972516773975998555269787930311658127446343921140043043176165836449102631845535019286744963120734410232430707114679579223296474215951229800332459875853037330883692346332242062914750404507192462901238392868499991240954288384442049");
-		String chiffreText = "12302552406977273930065439138838967203995032241714751614119290361982642066906515068510088545534307858844348347449786308731353961971050859372593393690645387024845476873192778761690959946280277775619476701212114903525769529799770945871326150627404853276653743351521585737077882218932045192337668885940408096496106264319975007131618909488242860856129705107818966989668525243208679164046621858590752138974238143882917264528128475152303981221364993395764329389736706655297771862682141592446104368278219498262105477339992476114165448767285554705600788614944846068939544301962000313412939300149442102018444122329245917565830,16341474788185962176518640399461202588775818479148137828217192889562344527522400721054457737545699910136456321517569621413824586049291541081138523773915807609464919382050457046978359384968814247208949129000244330792851525849186046193689432588920623180109586238009083547728225354775167664119533196380895196423035376506876627342007959782705487024923582339456390869487481646414238979723666207817516044655612050918935612006448075693806017533039188777988412636454960354183034380161067368537712394192382592508394012298684852473695050044495626745120258694254750757324521524597394939642108140527490904289533086137320258937984,11263529306877045936789598513129792464465084806781433366508342343019233205753343845058841031365375175632616230489471174640364205627656530298789089134766776676465558739728162192113479058731256173402440373533502929572896299274766628915985622195918302298757418422856948029616372915162367313022304089692273122860329276243792722043557231755302245014322226772696657638853317037094604538736140325443683979350116316372618196576937099153000005156374050614315934403990957906734440132201060454834870684260982032998817155995632722151335967064014744880324885605571942150320758607646609273039220138121177732598207039517058265546899,13608462333844604542786436080564393554452629841039653380570178093691952017782345607795350476859478424172034715236427394196714957033140244673933703183623621050933767301181762581174155862009660745241952395746068866518684021775167290479867599211915746402598025905700596649525758052923503567436216456685627295353686751611982858660811881828525018209933614107052504133511081776408756523367206006665146220796516977692006740804391560342289559977370805482058921449826634316753584704004024643276583994516235644880271460993506644517862535674888651226669140654445349120907002848366225767935907723947550784355861251981295989061248,6460594650335091010832041374346075294037972738219905800995808884582696176373559874366028411629781806453471469020490128055772813655439622025492822182221160753510402294216333929053217885939718904428846612214638890545842355498171019423803095976404199945993027519338032294039349894924970764752338075477582307951190817589857781579788483125521006484051735097899022710084479619168044036941952031199310671065642353883324577477840344496441265085751737061786052133009525983704824056092350077567808240093829235942759944716106110045502486802862265597510379809867530049184269452601916746240239984549822815314043745128298084835437,16341474788185962176518640399461202588775818479148137828217192889562344527522400721054457737545699910136456321517569621413824586049291541081138523773915807609464919382050457046978359384968814247208949129000244330792851525849186046193689432588920623180109586238009083547728225354775167664119533196380895196423035376506876627342007959782705487024923582339456390869487481646414238979723666207817516044655612050918935612006448075693806017533039188777988412636454960354183034380161067368537712394192382592508394012298684852473695050044495626745120258694254750757324521524597394939642108140527490904289533086137320258937984,8574687585533793061517289607218692687941521985356167009557626427490305608714123046537922139606714260365518407512376991084706109660765854236729702956980492077551386858716424362337132726896431700669479299291158309744632682008603535053867806711935634936460458513443972858092091315715131402610630631203049253121717688307091949589288197563576430178851772728417058139418121517737599002892187664926839290009941504666409526791305436840018100570102554067499421123239020797939223006227825102295501809710455039585255279784658843609469551939441753458466417751277223772400826936965261886618250444708777343862588318193562542982523,2878609762898137566794304114130501583918484152209731981078518656053929077512466063632773676174532432217987345857258330654032896067483473330610549235531319685121707303310401188048477988056673339065191911375681460506832531727885355208879685914314530831524091322636319123675929813801067836883225695767393725477457556550682764446626995183665547655004232093111628696009531852781529971184005238207565583848784004095731035683983055683337666891819677809793078232131411365648778666234328935704736459044674662042392967300394901513082381076855677921798398468170502142285993036551426215427213612741198770674080309518988133059150,10822186710816278828513901713633184553281058961566822246165604322141094957178062120011571742580078260741613769150998034159895616370626873712532443474436852385864908199320388319296453141281087895704521498041036859928905986197224836651520933937493270649262628464147392131915180053714230403613789359303112002825687525852795814175309930438448941544574129106981409406337814119072689510750003266417930708111297304582973504237224948651368678218130914063166067262075535716852576049336971642751918411455672056004589691095132397016236291399713991469333998458183021355249366166550428377987768236906077986845621545509961151190038,13608462333844604542786436080564393554452629841039653380570178093691952017782345607795350476859478424172034715236427394196714957033140244673933703183623621050933767301181762581174155862009660745241952395746068866518684021775167290479867599211915746402598025905700596649525758052923503567436216456685627295353686751611982858660811881828525018209933614107052504133511081776408756523367206006665146220796516977692006740804391560342289559977370805482058921449826634316753584704004024643276583994516235644880271460993506644517862535674888651226669140654445349120907002848366225767935907723947550784355861251981295989061248,2758061430832452371691472713298676003431783322818090189746203975081635906613403394631778883619251977558832620415083661909763575673726729037794158675767264819576629289171041278876087365588060102050587267808922539691767769566710797364890513753621234606402090635202677565906061607536051308181413242657969179038374017317163633779407823065698550596456312944022916301708854519118105008383118802627547169821176997818073701440435453090988966562892136780299587644421464776840910183134025422199433060433022618812533378130655099699537359351590369869931178411486903221062391691403338338696793686549925709528645507792525992705723,1598872275607504554917532261861132752393660959084558556132502327021656862851539460754619983390137238191131061025717885702802430999602631185652433241758843240282955612665975879087610627888655220184985257825910382060560015110739654857598835741167158960645742166932819541956097839343631086049778022372576891446459985545889761897247201415790080176440531338724383987117884870587688250279345272242997369534177843894858403449623520533322231399647654286728593247094073596706786233192473574988093892168506194228125413710368515153743032019322443799670954056594512926370681521734345958744613660332511213672134967796440691253685,2878609762898137566794304114130501583918484152209731981078518656053929077512466063632773676174532432217987345857258330654032896067483473330610549235531319685121707303310401188048477988056673339065191911375681460506832531727885355208879685914314530831524091322636319123675929813801067836883225695767393725477457556550682764446626995183665547655004232093111628696009531852781529971184005238207565583848784004095731035683983055683337666891819677809793078232131411365648778666234328935704736459044674662042392967300394901513082381076855677921798398468170502142285993036551426215427213612741198770674080309518988133059150,13608462333844604542786436080564393554452629841039653380570178093691952017782345607795350476859478424172034715236427394196714957033140244673933703183623621050933767301181762581174155862009660745241952395746068866518684021775167290479867599211915746402598025905700596649525758052923503567436216456685627295353686751611982858660811881828525018209933614107052504133511081776408756523367206006665146220796516977692006740804391560342289559977370805482058921449826634316753584704004024643276583994516235644880271460993506644517862535674888651226669140654445349120907002848366225767935907723947550784355861251981295989061248,12606024115830875251209458787815511599024657092818936024644992030249511035871340376005714754210325280929707979462621002220626144931636084878847552934381264083368260648952138898455891186333294744769653303426433018871805904171485783029025958583754998265285398214072922858659229164326661129876174933845153597731100213125186816796405069085109980308229990219667727195479988461585011421608059765146354920187063074461618346267539825498703776033444791554822916615635631637560515869343614971798258429487788556486633743140573497405973834378782551915610443019201514977031135340222905001839416013599895255396951847418362518483878,3068301564903762008367562474370711107756632637984836462363191103766550920093655324845046066152999569415564118479589112793195422789148164194720415127693595328342508285630305892577512196623563670997781919318879468100453036780199001016811214238968451107205667400685942729974796851228090277700545187195947578708288404252239907438391981165390472350431062878783565474202053184994579555133754310894693262035211925202444077593957521462792021044176579178075871023869544199980098027610828194253132403645759372348447769314196157761295236718568255976672426857296966117110737719602594901410086476815593138189746011605543920441420,8744529382761044513028121965930235455130117349102727457617207738602190348793628609869638797327328575309857800004583806603948976488501260159237771095385623828732356440114456408136430052047077462917524194049345509957194291401203953204599453516232627844559985978525910736485605334644753058350248368418709391666603080783597784491827067944512267146044748034996049179251449326908268654601256209201450650197730884659170634276674245287881412214529472154115470674335546311576612966934268894211462051783260894439662275212946742870874941646086586096741503278190957893447709282851158979491764966033525688679453542309801526190689,13608462333844604542786436080564393554452629841039653380570178093691952017782345607795350476859478424172034715236427394196714957033140244673933703183623621050933767301181762581174155862009660745241952395746068866518684021775167290479867599211915746402598025905700596649525758052923503567436216456685627295353686751611982858660811881828525018209933614107052504133511081776408756523367206006665146220796516977692006740804391560342289559977370805482058921449826634316753584704004024643276583994516235644880271460993506644517862535674888651226669140654445349120907002848366225767935907723947550784355861251981295989061248,12606024115830875251209458787815511599024657092818936024644992030249511035871340376005714754210325280929707979462621002220626144931636084878847552934381264083368260648952138898455891186333294744769653303426433018871805904171485783029025958583754998265285398214072922858659229164326661129876174933845153597731100213125186816796405069085109980308229990219667727195479988461585011421608059765146354920187063074461618346267539825498703776033444791554822916615635631637560515869343614971798258429487788556486633743140573497405973834378782551915610443019201514977031135340222905001839416013599895255396951847418362518483878,2878609762898137566794304114130501583918484152209731981078518656053929077512466063632773676174532432217987345857258330654032896067483473330610549235531319685121707303310401188048477988056673339065191911375681460506832531727885355208879685914314530831524091322636319123675929813801067836883225695767393725477457556550682764446626995183665547655004232093111628696009531852781529971184005238207565583848784004095731035683983055683337666891819677809793078232131411365648778666234328935704736459044674662042392967300394901513082381076855677921798398468170502142285993036551426215427213612741198770674080309518988133059150,6931539322581476100438709648410817845937301215033453568676241897693181975882928543422811093548776773881763650066997338346617321454940916285340749788482720231501669332270805548193464977034187942930540662932084692463890714519867425602435426321605335070479032104159167422714773392973021346638866607454176108293036870432434399815474557110556476275917403093566922473846759573833663393860097539545748115496675650634852295351349026800741370938094612281191841000281643443526582879281592005590467007338354253516655538948447081759941187699794489372591973306859907645850226461810605668880560840555838099713891126151058230291642,16341474788185962176518640399461202588775818479148137828217192889562344527522400721054457737545699910136456321517569621413824586049291541081138523773915807609464919382050457046978359384968814247208949129000244330792851525849186046193689432588920623180109586238009083547728225354775167664119533196380895196423035376506876627342007959782705487024923582339456390869487481646414238979723666207817516044655612050918935612006448075693806017533039188777988412636454960354183034380161067368537712394192382592508394012298684852473695050044495626745120258694254750757324521524597394939642108140527490904289533086137320258937984,9163379994849965475473213522896392064872782296546974121477189796744881678971348226896076407250514682570473310520402063709065401757054985788032561276311345586546850068151693264940008584918017260874687772411270913250705296100078257714497962971557673399700052364238296767418708939919951569993545186528394631786320835350693989557129162361387467856434580880893919492121572114804173483132473085291414214036874562779924557805106165883879407482745116993973845389751291381299418434363370976202208382057519501277592296479309980493513253632142418940900400837482303245537329545758617309262128674760657385831046966041796691332116,6460594650335091010832041374346075294037972738219905800995808884582696176373559874366028411629781806453471469020490128055772813655439622025492822182221160753510402294216333929053217885939718904428846612214638890545842355498171019423803095976404199945993027519338032294039349894924970764752338075477582307951190817589857781579788483125521006484051735097899022710084479619168044036941952031199310671065642353883324577477840344496441265085751737061786052133009525983704824056092350077567808240093829235942759944716106110045502486802862265597510379809867530049184269452601916746240239984549822815314043745128298084835437,8744529382761044513028121965930235455130117349102727457617207738602190348793628609869638797327328575309857800004583806603948976488501260159237771095385623828732356440114456408136430052047077462917524194049345509957194291401203953204599453516232627844559985978525910736485605334644753058350248368418709391666603080783597784491827067944512267146044748034996049179251449326908268654601256209201450650197730884659170634276674245287881412214529472154115470674335546311576612966934268894211462051783260894439662275212946742870874941646086586096741503278190957893447709282851158979491764966033525688679453542309801526190689,15367064846865234252191844930095295823985468888259018836334197781696982814384547140562558839229505132283623991058754150969660558244819564763514537686757258862238967006786226581039390663367155969011798897519101321916825848513742189539496559470597530857627298998685984076586965440588402647800215484804450616994866641210666666080308790204954835293020732275735811205459540373826968080652452932514503292036321536452155413104476196807745962919315202132734721880472907210894605013014846677203304040881219890813959987402839975925515451691152674768790812035895941165962006410955915260947719396510448510215756787315598329046679,6111586929544912821863732709295884197138303279654536712645733427962110460081319717191290958204193752961705865283017695228635413727253402183841391942436872661613121044033141517084006330367908684452238627227927899242136709132766140206610879336678931243529184794799294958085829820472221664774889957064517734666795391501770345463329358838014707716957084163383980313487115296818537614278251322474294228321339973080328514403132324325326137344750219593225032489207472756338839259382152549302629589669866479229131271288295467965838179717973258175476624252398501622167940517370777254558846857614848964149735524088626381740924,6111586929544912821863732709295884197138303279654536712645733427962110460081319717191290958204193752961705865283017695228635413727253402183841391942436872661613121044033141517084006330367908684452238627227927899242136709132766140206610879336678931243529184794799294958085829820472221664774889957064517734666795391501770345463329358838014707716957084163383980313487115296818537614278251322474294228321339973080328514403132324325326137344750219593225032489207472756338839259382152549302629589669866479229131271288295467965838179717973258175476624252398501622167940517370777254558846857614848964149735524088626381740924,6111586929544912821863732709295884197138303279654536712645733427962110460081319717191290958204193752961705865283017695228635413727253402183841391942436872661613121044033141517084006330367908684452238627227927899242136709132766140206610879336678931243529184794799294958085829820472221664774889957064517734666795391501770345463329358838014707716957084163383980313487115296818537614278251322474294228321339973080328514403132324325326137344750219593225032489207472756338839259382152549302629589669866479229131271288295467965838179717973258175476624252398501622167940517370777254558846857614848964149735524088626381740924,6111586929544912821863732709295884197138303279654536712645733427962110460081319717191290958204193752961705865283017695228635413727253402183841391942436872661613121044033141517084006330367908684452238627227927899242136709132766140206610879336678931243529184794799294958085829820472221664774889957064517734666795391501770345463329358838014707716957084163383980313487115296818537614278251322474294228321339973080328514403132324325326137344750219593225032489207472756338839259382152549302629589669866479229131271288295467965838179717973258175476624252398501622167940517370777254558846857614848964149735524088626381740924,6111586929544912821863732709295884197138303279654536712645733427962110460081319717191290958204193752961705865283017695228635413727253402183841391942436872661613121044033141517084006330367908684452238627227927899242136709132766140206610879336678931243529184794799294958085829820472221664774889957064517734666795391501770345463329358838014707716957084163383980313487115296818537614278251322474294228321339973080328514403132324325326137344750219593225032489207472756338839259382152549302629589669866479229131271288295467965838179717973258175476624252398501622167940517370777254558846857614848964149735524088626381740924";
-		String decrypt = app.decrypt(new Tuple(first, second), chiffreText);
-		Assert.assertEquals("Das haben Sie gut gemacht!\t\t\t\t\t", decrypt);
-	}
-
 }
